Fix final score
When the score is decided on the final play, it seems the line-score ignores the points for the FINAL SCORE, though not for the 4th quarter. I reqrote the scoreboard part so that the final score is not pulled from the data file but is rather calculated by adding the quarter scores.

Has **not** been tested in an OT situation.